% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_whit.R
\docType{data}
\name{palette_whit}
\alias{palette_whit}
\title{Whit's [ggplot2] theme}
\format{
A \code{list}.
}
\usage{
palette_whit
}
\description{
Palette data for whit
}
\details{
Data used by the palettes in the whitrthemes package.
}
\keyword{datasets}
