% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_print.R
\name{theme_print}
\alias{theme_print}
\title{A \link{ggplot2} theme for Whit}
\usage{
theme_print(
  base_size = 8.5,
  base_family = "Sans",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_print} provides a \link{ggplot2} formatted theme.
}
\details{
Note: adapted from \link{urbnthemes}
}
