% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{whit_color_pal}
\alias{whit_color_pal}
\title{Whit's [ggplot2] theme}
\usage{
whit_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes that whit likes.
}
