% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_whit_main}
\alias{palette_whit_main}
\alias{palette_whit_diverging}
\alias{palette_whit_quintile}
\alias{palette_whit_politics}
\alias{palette_whit_cyan}
\alias{palette_whit_gray}
\alias{palette_whit_yellow}
\alias{palette_whit_magenta}
\alias{palette_whit_green}
\alias{palette_whit_spacegray}
\alias{palette_whit_red}
\title{whit's palettes}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 7.

An object of class \code{character} of length 5.

An object of class \code{character} of length 2.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\source{
Learn UI Design Palette Generator
 \url{https://www.learnui.design/tools/data-color-picker.html#palette}
}
\usage{
palette_whit_main

palette_whit_diverging

palette_whit_quintile

palette_whit_politics

palette_whit_cyan

palette_whit_gray

palette_whit_yellow

palette_whit_magenta

palette_whit_green

palette_whit_spacegray

palette_whit_red
}
\description{
Vectors with hex-color codes
}
\concept{whit palettes}
\keyword{datasets}
