% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_plot.R
\name{whit_plot}
\alias{whit_plot}
\title{urbn_plot}
\usage{
whit_plot(..., heights = 1)
}
\arguments{
\item{...}{urbn plot objects or grobs}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
Combine elements from \code{whit_title},
\code{whit_y_title}, \code{get_legend}, \code{remove_legend},
\code{whit_notes}, \code{whit_source}, and \code{ggplot2} into
one formatted plot.
}
