% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_save.R
\name{whit_save}
\alias{whit_save}
\title{Save ggplot2 plots in standard sizes}
\usage{
whit_save(
  filename,
  plot = ggplot2::last_plot(),
  size = "medium",
  dpi = 300,
  height = NULL
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{size}{Size of plot. Must be one of "small" (3.25 x 2 inches),
"medium" (6.5 x 4 inches), or "large" (9 x 6.5 inches).}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
}
\description{
This is a function to save standardized images.
}
