% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_source.R
\name{whit_source}
\alias{whit_source}
\title{whit_source}
\usage{
whit_source(text, size = 12, width = 120, plural = FALSE)
}
\arguments{
\item{text}{character string for a source}

\item{size}{font size for the source}

\item{width}{a number of characters to allow before a character return}

\item{plural}{If `TRUE`, will change "Sources:" to "Sources:"}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a sources section to the bottom of a plot created with \code{whit_plot()}.
}
