% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_title.R
\name{whit_title}
\alias{whit_title}
\title{whit_title}
\usage{
whit_title(string, size = 17.5, width = 75)
}
\arguments{
\item{string}{character string for a title}

\item{size}{font size for the title}

\item{width}{a number of characters to allow before a character return}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a title to a plot created with \code{whit_plot()}.
}
