% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_y_title.R
\name{whit_y_title}
\alias{whit_y_title}
\title{whit_y_title}
\usage{
whit_y_title(string, size = 16)
}
\arguments{
\item{string}{character string for a y-axis title}

\item{size}{font size for the y-axis title}
}
\value{
a grob formatted for a y-axis title in a ggplot
}
\description{
Add a horizontal y axis title to a plot created with \code{urbn_plot()}.
}
